Omadeb 1.0.0-rc1

+++ applications/install/docker.sh
@@ -1,20 +1,15 @@
 #!/bin/bash
 
-# Install Docker
-# Add Docker's official GPG key:
-sudo apt-get update
-sudo apt-get install ca-certificates curl
-sudo install -m 0755 -d /etc/apt/keyrings
-sudo curl -fsSL https://download.docker.com/linux/debian/gpg -o /etc/apt/keyrings/docker.asc
-sudo chmod a+r /etc/apt/keyrings/docker.asc
+# Add the official Docker repo
+if [ ! -f /etc/apt/sources.list.d/docker.list ]; then
+    [ -f /etc/apt/keyrings/docker.asc ] && sudo rm /etc/apt/keyrings/docker.asc
+    sudo install -m 0755 -d /etc/apt/keyrings
+    sudo wget -qO /etc/apt/keyrings/docker.asc https://download.docker.com/linux/debian/gpg
+    sudo chmod a+r /etc/apt/keyrings/docker.asc
+    echo "deb [arch=$(dpkg --print-architecture) signed-by=/etc/apt/keyrings/docker.asc] https://download.docker.com/linux/debian $(. /etc/os-release && echo "$VERSION_CODENAME") stable" | sudo tee /etc/apt/sources.list.d/docker.list >/dev/null
+fi
 
-# Add the repository to Apt sources:
-echo \
-  "deb [arch=$(dpkg --print-architecture) signed-by=/etc/apt/keyrings/docker.asc] https://download.docker.com/linux/debian \
-  $(. /etc/os-release && echo "$VERSION_CODENAME") stable" | \
-  sudo tee /etc/apt/sources.list.d/docker.list > /dev/null
 sudo apt update
-
 sudo apt install -y docker-ce docker-ce-cli containerd.io docker-buildx-plugin docker-compose-plugin docker-ce-rootless-extras
 
 # Configure Docker daemon:
@@ -30,8 +25,13 @@ sudo tee /etc/docker/daemon.json >/dev/null <<'EOF'
 }
 EOF
 
+# Expose systemd-resolved to our Docker network
+sudo mkdir -p /etc/systemd/resolved.conf.d
+echo -e '[Resolve]\nDNSStubListenerExtra=172.17.0.1' | sudo tee /etc/systemd/resolved.conf.d/20-docker-dns.conf >/dev/null
+sudo systemctl restart systemd-resolved
+
 # Start Docker automatically
-sudo systemctl enable docker
+sudo systemctl enable docker.socket
 
 # Give this user privileged Docker access
 sudo usermod -aG docker ${USER}
